Dev UI 2 - merge build steps and build items pages (not visible yet)
...into the build metrics page with tabs
